1.加入计算jsp文件中的Java有效代码 2.修复小bug

+++ src/com/wsy/CCGo.java
@@ -26,7 +26,14 @@ public class CCGo {
 			System.out.println("Usage:java -jar CodeCounter.jar \"C:/Users/diazrael/Desktop/lib\" ");
 			return;
 		}
+		System.out.println("注意：本工具仅计算有效java代码，可用于报价的辅助工具。");
+		System.out.println("以下情况算作有效代码：");
+		System.out.println("1.Java文件中的有效Java代码（可排除注释）。"
+				+ "2.jsp文件中的Java代码（可排除注释）。"
+				+ "3.jar包中反编译出来的java代码（此项过程中可选）。");
+		
 		JavaTotal.javaPath = args[0];
+		//判断是否计算jar包中的行数
 		String jarflag = readDataFromConsole("Whether include jars?(Y/N)");  
 		if("Y".equalsIgnoreCase(jarflag)){
 			// 解压缩Jar
@@ -37,13 +44,18 @@ public class CCGo {
 		
 		// 调用javacounter计数
 		JavaTotal.runJavaTotal();
-		String delflag = readDataFromConsole("Whether to delete the directory compiled?(Y/N)");  
-		if("Y".equalsIgnoreCase(delflag)){
-			for (String dir : UnZip.classFileList) {
-				UnZip.deleteDir(new File(dir));
+		//是否删除反编译之后的文件夹
+		if("Y".equalsIgnoreCase(jarflag)){
+			String delflag = readDataFromConsole("Whether to delete the directory compiled?(Y/N)");  
+			if("Y".equalsIgnoreCase(delflag)){
+				for (String dir : UnZip.classFileList) {
+					UnZip.deleteDir(new File(dir));
+				}
 			}
 		}
+		
 		long end = System.nanoTime();
+		System.out.println("Result saved in ./javaFileCount.txt.");
 		System.out.print("cost: " + (end - start) / 1e9 + " seconds");
 	}
